refactor(compact): widen num_threads to u64 for ABI uniformity
All other numeric fields in LanceCompactionOptions are u64; switching
num_threads to match removes the 4-byte padding hole and lets the resolve
path use u64_to_usize uniformly (overflow-checked) rather than an as-cast.

+++ include/lance/lance.h
@@ -408,7 +408,7 @@ typedef struct LanceCompactionOptions {
     uint64_t max_bytes_per_file;
     /* Compute parallelism for compaction tasks. 0 = default
        (number of compute-intensive CPUs). */
-    uint32_t num_threads;
+    uint64_t num_threads;
     /* Scanner batch size for reading input fragments. 0 = default. */
     uint64_t batch_size;
 } LanceCompactionOptions;
